+++ src/bbcollapsible/index.tsx
@@ -1,5 +1,6 @@
 import classnames from 'classnames';
 import React, { useState, Children, cloneElement } from 'react';
+import BBButton from '../bbbutton';
 import BBCard from '../bbcard';
 import styles from './styles.module.scss';
 import type { IPropsBBCard } from '../bbcard';